[CI] Add no-infra-wheels jobs
Since Kolla is run in the wild against PyPI mirrors which do not
include the OpenDev-provided wheels, it might be nice to be able
to test these scenarios in the CI.

These jobs are added to the experimental queue so that they do
not consume resources when not needed.
Any Gerrit user may run them now to verify whether the change
passes them.

Change-Id: I9c33eeaf03be037651f82c831f0e07ab9a27752a