Fix Masakari dashboard policy file location
The horizon image copies the masakari dashboard's policy file to Python
site packages, but it should go to /etc/openstack-dashboard. This allows
the dashboard to be seen by non-admins, although it fails to load.

We get an error like the following in horizon.log:

No policy rules for service 'instance-ha' in
/etc/openstack-dashboard/masakari_policy.json

This change fixes the issue.

Change-Id: I8ede183c76a830de06ce6524dc3f6f6944b182c1
Closes-Bug: #1894240