CI: revert to public package mirrors after build
When building images in CI, we use local package mirrors to reduce
external network traffic. These mirrors are not necessarily reliable nor
are they guaranteed to be accessible outside of the CI environment
(although many are).

While it is an antipattern to install packages in containers at runtime,
the Bifrost container does just that. If the mirrors configured in the
image are unavailable, Bifrost deployment can fail.

This change fixes the issue by reverting to public package mirrors as a
final step of the image build process.

Closes-Bug: #1902101

Change-Id: I2592a736206eaec811290e9fbdbf2540c0518ffe
(cherry picked from commit 0fc464332c789ce05e9d6442726ada84ce7fc65a)