CI: Improve publish playbook
Currently we use a for loop in shell that performs push of images
to Docker Hub - that doesn't really seem to fail properly on push
errors (e.g. related to new Docker Hub limits).

This change introduces usage of Ansible docker_* modules for doing the
same job and improves logging (so it's easier to find out which image
failed to publish).

Change-Id: I083ce70772edd1abb873ad43f6beb73aae01383f