[CI] Slim down
Move all extra jobs to the experimental pipeline.
Cores should know when to test these extras.
This is to reduce the load when a fix is needed in the base image
or across all images.
Otherwise, due to how CI works, we incur the cost of running
all these jobs on each recheck.

Change-Id: Ia4aecd2fbe70d1760e0ddc401d06510914dbd1fb