Upgrade the Monasca Fluentd output plugin
From Victoria onwards a conflict is seen between the rest-client
library used by the Monasca Fluentd output plugin and the openssl
library. Specifically, a more recent release of openssl has dropped
a default which is referenced by the 1.8 release of the rest-client
library. In this commit we bump the Fluentd output plugin release
to pull in a more recent version of the rest-client library which
can handle this change.

Change-Id: Iae5402ea40c9f856b1787fdf82cc2389ffe0be03
Closes-Bug: 1910382