Horizon: gentler `-o nounset` handling
When running with `-o nounset` since [1], the Horizon became quite
fragile to run as it started requiring all ENABLE_* environment
variables to be set upfront. Normally they are - via kolla-ansible.
However, when working with it outside of kolla-ansible or removing
services (like it was the case 3 times during the Wallaby cycle),
it creates needless issues (like having to wait for images to get
published for kolla-ansible gate or users bumping into irrelevant
incompatibilities [2]).

This patch makes sure all ENABLE_* environment variables default
to 'no' and are no longer required to be set when there is no
need to set them to 'yes'.

[1] 032804e5a0ddf89f7726880d1b57dd492d7d5c07
[2] https://bugs.launchpad.net/kolla/+bug/1911141

Change-Id: I644e072a699dccd8f32a24e484ff6dab7b9b449d