base: check for compatible distro using just sh
We allow users to use own image as base one and then we check is it
proper distribution version one. And this check is done in different way
for CentOS and different for Debian/Ubuntu systems.

This patchset uses PRETTY_NAME from /etc/os-release for check. And it is
done before anything gets installed so if wrong distribution is used
then build process breaks as soon as possible.

Change-Id: I22db3cd4cb608332261824626e3b8768821b8dd0