monasca-thresh: Allow topology check and removal in storm
Patch adds a script in the monasca-thresh image that can be used
to check if a topology exists in Storm, and optionally kill it.

This is part of a bug in kolla-ansible where topologies were
not submitted to Storm, but run locally.  This patch includes
a topology check script enabled by KOLLA_BOOTSTRAP which will exit
kolla_start if the topology exists, and optionally enables topology
removal (to allow replacement) enabled by TOPOLOGY_REPLACE.

Topology names and various timeouts may be customized.  If the
new env variables are not set, existing behavior is unchanged.

Partial-Bug: #1808805
Change-Id: If8f0730031435dda4235b7f2d2c23e5f5f767f87