Pin td-agent to 4.0.* to fix missing logs
The recent inclusion of Fluentd 1.12.x in Centos/Ubuntu packages
causes a significant number of logs to fail to appear in the
Elasticsearch or Monasca output plugins. The logs appear to
sit in a local buffer, and are only flushed when Fluentd is
restarted. The same issue persists in the most recent release
(1.13.0), so for now we pin the td-agent package to 4.0.* which
should restrict Fluentd to the 1.11.x release where we don't
see this issue.

Backport to Victoria.

Change-Id: Iefcdd3100b7e3c5320bc5f1286a18251bdeab885
Closes-Bug: #1930867