Fix missing pacemaker-cli-utils in Debian hacluster images
Hacluster-pacemaker images need to have pacemaker cli utils
available in all distros and they are provided by pacemaker-cli-utils
package in Debuntu systems. Due to a different dependency chain
this package is not installed in Debian. This package is also
*recommended* dependency of pacemaker and pacemaker-remote in both
Debian/Ubuntu systems, but unfortunatelly it is not installed
on Debian via APT by default.

This patch adding this package to list of package to be installed
on Debian/Ubuntu systems for both hacluster-pacemaker (required)
and hacluster-pacemaker-remote (to ease with debugging).

Closes-Bug: #1934788
Change-Id: I055086ab4af37878e81844f1055bc48d84ae9945
(cherry picked from commit 535c861d99aa0c07ea130d1f30cabb6af3db7ab7)