Revert "Update code to Jinja 3.0 style"
This reverts commit 383364e671fb82e3c0ddca37a4690a1a33be3b54.

Reason for revert: Xena stay in jinja 3.0.1 which still provides contextfilter. In RDO, jinja is consumed from CentOS repo which still provides version 2.11, so this change is causing problems.

I'm proposing to revert this change and retake it Y release when jinja is updated to 3.1 which will not provide contextfilter.

Change-Id: I19c2a15fb0917d2a81daf706a4e02fd377e53b3d