logstash: pin to 7.9.* for Elasticsearch OSS compatibility
The latest versions of logstash 7.x do not support using the OSS version
of Elasticsearch, and fail to push logs if an OSS version is used. This
prevents Monasca log API from working correctly.

This change fixes the issue by pinning logstash-oss to 7.9.x, the most
recent available package in the repo that still supports Elasticsearch
OSS.

NOTE: logstash 7.11 introduced the breaking change, however 7.10 is
unavailable in the package repos.

Change-Id: I4ec4eb837c98f57964488769ab82ba8213ad582c
Closes-Bug: #1941754