Ensure nvme-cli is present in nova-compute images
The nvme-cli package is required to get the nvme command executed by
os-brick, which is used by nova-compute.

We don't need to explicitely install it for centos binary images, as it
is required by os-brick which is required by openstack-nova-common, but
all other types of images were missing it.

Change-Id: I754939da7636c57d2a8d5b83debb5d8a58e38432
Closes-Bug: #1953509
(cherry picked from commit 30eddf72f6d5380bb6078481581e47b0eec84982)