Clean up fluentd image recipe
Since Xena, there is no arch supported other than x86_64 and
aarch64.

Additionally, the conditional for x86_64 Debian does not seem to be
needed now with current td-agent.

Moreover, drop two plugins that are relevant only to fluentd 0.12,
one of which (parser) causes fluentd 0.12 (!) to be installed.
Modern fluentd has both built-in.

Finally, drop the labels as they are no longer useful.

Depends-On: https://review.opendev.org/c/openstack/kolla-ansible/+/823093
Depends-On: https://review.opendev.org/c/openstack/kolla-ansible/+/823094
Change-Id: I7326cd8fcfa459dedcb6404e5287f0f5aeeed6cb