Remove missing collectd packages
Since CentOS Linux 8 repository retirement from
CentOS mirrors - we need to use CentOS Stream 8
version of Opstools repo.

Listed below packages are missing from the repo,
and according to the maintainer dpdk_telemetry is
not going to show up at all (it's basically dead).

The rest of the packages might show up at a later
stage - but it's not a priority for the maintainer.

- collectd-dpdk_telemetry
- collectd-libpod-stats
- collectd-sensubility
- python3-sqlalchemy-collectd

Change-Id: I5f446fc276e245c1e6974ceeb6ff3dccd84ee502