Clarify documentation about source override
The current documentation about source override is ambiguous and let
users presume that you should always use "-base" after the project name
you want to override source location.

This assumption is wrong, as kolla look for a section who match exactly
the name of the image. Images who don't have a "-base", like Horizon,
are really hard to override for a user.

I propose to put an example without a "-base" and also add a disclaimer
in the documentation about that.

Change-Id: I308a352975f11c29afb3f2ea1c22212062066b4b