Use jinja2.pass_context instead of contextfilter
The contextfilter decorator was deprecated in jinja2 3.0.0, and has been
dropped in 3.1.0. This results in the following warning, and failed
attempts to use filters:

    [WARNING]: Skipping plugin (filters.py) as it seems to be invalid:
    module 'jinja2' has no attribute 'contextfilter'

This change switches to use the pass_context decorator. The minimum
version of Jinja2 is raised to 3 to ensure pass_context is present.

This change has been updated to also support Jinja2 2.x releases,
since the Wallaby upper constraints specify 2.11.3. In practice, most
users will not use UC to install kolla.

CoAuthored-by: Mark Goddard <mark@stackhpc.com>

Change-Id: I5efab66e487e06abd1a56af97d7e7caa1ebc880d