Revert "CI: add templated Dockerfiles to build logs"
This reverts commit 09c890b12463a4b2c8767231189e7016f337fbc3.

Reason for revert: That change caused all Kolla's work items to be
transferred. This includes tarballs, their extracted contents, cloned
git repositories, etc. to also be transferred to swift which takes
extra time, pollutes logs and decreases the stability of jobs slightly
(the mechanism is not designed for this kind of load).
The collection of templated Dockerfiles in CI has to be reimplemented
if it is wanted.

Change-Id: Ideaf66ce18c993802f4d20f490caa47cec0dde5e
(cherry picked from commit f33ef03977536ecb29ea464b9e557a6caacf4345)