Fix redis-sentinel containers always reporting changed
The CONFIG REWRITE command rewrites the redis.conf
file the server was started with, applying the minimal
changes needed to make it reflect the configuration
currently used by the server, which may be different
compared to the original one because of the use of
the CONFIG SET command.

https://redis.io/commands/config-rewrite/

Because of above behaviour it's needed to
hack kolla's CMD.

Without this hack container's
/usr/local/bin/kolla_set_configs --check
is always reporting changed state even if
config is not changed.

Therefore redis containers are always restarted
by kolla-ansible. This patch is fixing this.

Change-Id: I8245b9e753fa4175dacfa679a078f63876b0d327
Partial-Bug: #1967664