Fix local sources of git repositories
This is I2cbf1f539880d512aa223c3ef3a4b19ee18854ac extended to fix
the case when a git repository is used with a git repo.
This is probably a rarer use case but, still, we use it in CI
for in-review changes testing.

Change-Id: I77b0dcd2e9dfd8ea8390a471b80c8954b67ef91b
(cherry picked from commit b888f68daf6cb3ed9782fe85045858b6a04d4911)