[bifrost] Force Bifrost to use the correct u-c
Following up on [1], Bifrost also needs to use correct u-c and,
surprise-surprise, this is set via a third variable, this time
environment one for the bash script. Oh well. Let's do it.
This has the added benefit of not requiring additional
download of u-c as it reuses the already existing one.

[1] https://review.opendev.org/c/openstack/kolla/+/842276

Change-Id: I2cc79c476d089977b57006dd4ce2e9a8c6206aa1