Update stable release generation doc
Wallaby has been moved to EM state, so it will receive no more releases.
Also, being optimistic, zed is already being added to the list of stable
branches, so we don't forget it when we actually cut that branch.

Add a small note about this being referenced once per month in regular
Kolla meetings. Also update the example link to a patch that actually
matches the description.

Change-Id: Ie3b4e4461dbfe35808250a929b3e95f880242aed