Rework horizon image to support local_settings.d
The local_settings.py config is rendered by kolla-ansible
orchestration, which is not a good approach because upstream
local_settings.py can be changed anytime and that can be a problem,
as it was shown recently with the new version of horizon.

Fortunately, horizon supports local_settings overrides in
local_settings.d directory and moreover it's preffered
way how to configure horizon as per doc [1].

This patch just change the structure of files to support it.

[1] https://docs.openstack.org/horizon/latest/configuration/settings.html

Change-Id: Ib0c060adffe5287d786ba9247c6b962732cdb5e0