Install ironic-inspector in bifrost
Bifrost has moved to use the native in-band inspection support in
Ironic[1]. Some features are missing (such as introspection rules), so
it may still be desirable to use ironic-inspector for these use cases.
It also requires changes to the inspection interface of the ironic nodes
(ironic-inspector -> agent), so it also desirable to support
ironic-inspector for backwards compatability reasons. We should consider
dropping ironic-inspector once the native interface reaches feature
parity.

This will install ironic-inspector into the image, but will still
require configuration in bifrost.yml to use it.

[1] https://github.com/openstack/bifrost/commit/dcbf1d98b92236be7e4b9d89169a66892127e63b

Change-Id: I3cc8455647339c67ba51620d5aff8a3acbf58dab
Partial-Bug: #2054685