Fix state checking via kolla_set_configs
Kolla checks the status of its configurations
using kolla_set_configs --check. However, this
doesn't work for the command block in the config.json
file. This means that if only the command value changes,
but not the config_files block, which is common practice
in Kolla-Ansible btw, the container will not report the
changes.

This is undesirable and important to fix, especially
considering the planned rework of notifiers in
Kolla-Ansible [1].

[1] https://review.opendev.org/c/openstack/kolla-ansible/+/773243/14

Closes-Bug: #2080861
Change-Id: I2a290da38ea34b05ce3da8fb8b39b6252bf2da47