Added systemd-standalone-sysusers package for Debuntu
This package provides a standalone sysusers binary with
minimal dependencies, designed for use in non-systemd
environments.

This package simply ensures that all packages that want
to install systemd as dependency will have to settle for
this one because the systemd package breaks
systemd-standalone-sysusers, as it is an alternative.

This backport is necessary to fix failing Opensearch
builds while backporting Noble support to 2024.1.

[1] https://packages.debian.org/bookworm/systemd-standalone-sysusers
[2] https://packages.ubuntu.com/noble/systemd-standalone-sysusers

Needed-by: Ie12c43dc1de122c203edc28181366457ccd67834
Change-Id: I3ba85069c265844bd2f8b5e6d2b14f32e4898cfd
(cherry picked from commit 55168f909d3c95b6dfc386d3fbee25fd20a7cae8)