[letsencrypt] Rename script for updating certificates
This patch renames the update-haproxy-cert script to
update-loadbalancer-cert to make the name more generic.
The change is necessary because support for ProxySQL
will be introduced in the next patch, requiring a name
that reflects broader applicability.

To keep the diff in the following patch clean and easier
to review, this patch focuses exclusively on the rename.

Change-Id: I15f9b5cf3976f4298cdc7da4530f2d7d09614e88