neutron: Remove networking-generic-switch from u-c
networking-generic-switch was added to u-c [1] and it breaks our
install scheme.

Alternative is to remove networking-generic-switch from sources.py,
but that would be a user facing change and if we would pursue that
we would need to do the same for all neutron plugins.

using u-c removal macro to preserve previous approach of installing
n-g-s from a branch tarball compared to a pypi version

[1]: https://review.opendev.org/c/openstack/requirements/+/964059

Change-Id: Ifcee01f9415e9b9de8138b55cd676873a99c86aa
Signed-off-by: Michal Nasiadka <mnasiadka@gmail.com>